# artifacts/api-server/src/routes/stripe.ts
import { Router, type IRouter } from "express";
import {
  CancelStripePaymentIntentParams,
  CreateStripePaymentIntentBody,
  CreateStripePixIntentBody,
  CreateStripeSetupIntentBody,
  GetStripeContractPaymentsParams,
  GetStripePaymentIntentParams,
} from "@workspace/api-zod";
import { getStripeCredentials, getUncachableStripeClient } from "../lib/stripeClient";
import {
  getStripePaymentLinksForContract,
  upsertStripePaymentLink,
} from "../lib/stripeStorage";
import { requireAuth } from "../middleware/auth";

const router: IRouter = Router();
function asSingleParam(value: string | string[] | undefined): string {
  if (Array.isArray(value)) return value[0] ?? "";
  return value ?? "";
}

function normalizeAmountCents(body: { amountCents?: number; amount?: number }) {
  if (typeof body.amountCents === "number") {
    return Math.round(body.amountCents);
  }

  if (typeof body.amount === "number") {
    return Math.round(body.amount * 100);
  }

  return 0;
}

router.get("/stripe/config", async (_req, res) => {
  try {
    const credentials = await getStripeCredentials();

    res.json({
      configured: true,
      publishableKey: credentials.publishableKey ?? null,
      source: credentials.source,
    });
  } catch (error) {
    res.status(503).json({
      configured: false,
      publishableKey: null,
      error: error instanceof Error ? error.message : "Stripe is not configured.",
    });
  }
});

router.post("/stripe/payment-intents", requireAuth, async (req, res) => {
  try {
    const bodyParse = CreateStripePaymentIntentBody.safeParse(req.body);
    if (!bodyParse.success) {
      res.status(400).json({ error: "Invalid payload for create payment intent." });
      return;
    }
    const body = bodyParse.data;
    const contractId = body.contractId?.trim();
    const amountCents = normalizeAmountCents(body);
    const currency = (body.currency ?? "brl").toLowerCase();

    if (!contractId) {
      res.status(400).json({ error: "contractId is required." });
      return;
    }

    if (!Number.isFinite(amountCents) || amountCents <= 0) {
      res.status(400).json({ error: "amount or amountCents must be greater than zero." });
      return;
    }

    const stripe = await getUncachableStripeClient();
    let customerId: string | undefined;

    if (body.customerEmail) {
      const existing = await stripe.customers.list({
        email: body.customerEmail,
        limit: 1,
      });
      if (existing.data.length > 0) {
        customerId = existing.data[0].id;
      } else {
        const customer = await stripe.customers.create({
          email: body.customerEmail,
          name: body.customerName,
          metadata: {
            payer_profile_id: body.payerProfileId ?? "",
          },
        });
        customerId = customer.id;
      }
    }

    const paymentIntent = await stripe.paymentIntents.create({
      amount: amountCents,
      currency,
      customer: customerId,
      automatic_payment_methods: { enabled: true },
      metadata: {
        contract_id: contractId,
        payer_profile_id: body.payerProfileId ?? "",
        payee_profile_id: body.payeeProfileId ?? "",
        ...(body.metadata ?? {}),
      },
    });

    await upsertStripePaymentLink({
      contract_id: contractId,
      stripe_payment_intent_id: paymentIntent.id,
      amount_cents: amountCents,
      currency,
      status: paymentIntent.status,
      customer_id: customerId ?? null,
      payer_profile_id: body.payerProfileId ?? null,
      payee_profile_id: body.payeeProfileId ?? null,
      metadata: paymentIntent.metadata,
    });

    res.status(201).json({
      paymentIntentId: paymentIntent.id,
      clientSecret: paymentIntent.client_secret,
      status: paymentIntent.status,
      amountCents,
      currency,
    });
  } catch (error) {
    res.status(500).json({
      error: error instanceof Error ? error.message : "Failed to create Stripe PaymentIntent.",
    });
  }
});

router.get("/stripe/payment-intents/:paymentIntentId", requireAuth, async (req, res) => {
  try {
    const paramsParse = GetStripePaymentIntentParams.safeParse({
      paymentIntentId: asSingleParam(req.params.paymentIntentId),
    });
    if (!paramsParse.success) {
      res.status(400).json({ error: "Invalid paymentIntentId." });
      return;
    }
    const paymentIntentId = paramsParse.data.paymentIntentId;
    const stripe = await getUncachableStripeClient();
    const paymentIntent = await stripe.paymentIntents.retrieve(paymentIntentId);

    res.json({
      paymentIntentId: paymentIntent.id,
      status: paymentIntent.status,
      amountCents: paymentIntent.amount,
      currency: paymentIntent.currency,
      metadata: paymentIntent.metadata,
    });
  } catch (error) {
    res.status(500).json({
      error: error instanceof Error ? error.message : "Failed to retrieve Stripe PaymentIntent.",
    });
  }
});

router.post("/stripe/payment-intents/:paymentIntentId/cancel", requireAuth, async (req, res) => {
  try {
    const paramsParse = CancelStripePaymentIntentParams.safeParse({
      paymentIntentId: asSingleParam(req.params.paymentIntentId),
    });
    if (!paramsParse.success) {
      res.status(400).json({ error: "Invalid paymentIntentId." });
      return;
    }
    const paymentIntentId = paramsParse.data.paymentIntentId;
    const stripe = await getUncachableStripeClient();
    const paymentIntent = await stripe.paymentIntents.cancel(paymentIntentId);

    await upsertStripePaymentLink({
      contract_id: paymentIntent.metadata.contract_id ?? "unknown",
      stripe_payment_intent_id: paymentIntent.id,
      amount_cents: paymentIntent.amount,
      currency: paymentIntent.currency,
      status: paymentIntent.status,
      customer_id:
        typeof paymentIntent.customer === "string"
          ? paymentIntent.customer
          : paymentIntent.customer?.id ?? null,
      payer_profile_id: paymentIntent.metadata.payer_profile_id ?? null,
      payee_profile_id: paymentIntent.metadata.payee_profile_id ?? null,
      metadata: paymentIntent.metadata,
    });

    res.json({
      paymentIntentId: paymentIntent.id,
      status: paymentIntent.status,
    });
  } catch (error) {
    res.status(500).json({
      error: error instanceof Error ? error.message : "Failed to cancel Stripe PaymentIntent.",
    });
  }
});

router.get("/stripe/contracts/:contractId/payments", requireAuth, async (req, res) => {
  const paramsParse = GetStripeContractPaymentsParams.safeParse({
    contractId: asSingleParam(req.params.contractId),
  });
  if (!paramsParse.success) {
    res.status(400).json({ error: "Invalid contractId." });
    return;
  }
  const payments = await getStripePaymentLinksForContract(paramsParse.data.contractId);
  res.json({ data: payments });
});

router.post("/stripe/setup-intents", requireAuth, async (req, res) => {
  try {
    const bodyParse = CreateStripeSetupIntentBody.safeParse(req.body ?? {});
    if (!bodyParse.success) {
      res.status(400).json({ error: "Invalid payload for setup intent." });
      return;
    }
    const body = bodyParse.data;
    const stripe = await getUncachableStripeClient();
    let customerId: string | undefined;

    if (body.customerEmail) {
      const existing = await stripe.customers.list({ email: body.customerEmail, limit: 1 });
      if (existing.data.length > 0) {
        customerId = existing.data[0].id;
      } else {
        const customer = await stripe.customers.create({
          email: body.customerEmail,
          name: body.customerName,
          metadata: { payer_profile_id: body.payerProfileId ?? "" },
        });
        customerId = customer.id;
      }
    }

    const setupIntent = await stripe.setupIntents.create({
      customer: customerId,
      payment_method_types: ["card"],
    });

    res.status(201).json({
      setupIntentId: setupIntent.id,
      clientSecret: setupIntent.client_secret,
    });
  } catch (error) {
    res.status(500).json({
      error: error instanceof Error ? error.message : "Failed to create SetupIntent.",
    });
  }
});

router.post("/stripe/pix-intents", requireAuth, async (req, res) => {
  try {
    const bodyParse = CreateStripePixIntentBody.safeParse(req.body);
    if (!bodyParse.success) {
      res.status(400).json({ error: "Invalid payload for pix intent." });
      return;
    }
    const body = bodyParse.data;
    const contractId = body.contractId?.trim();
    const amountCents = normalizeAmountCents(body);

    if (!contractId) {
      res.status(400).json({ error: "contractId is required." });
      return;
    }

    if (!Number.isFinite(amountCents) || amountCents <= 0) {
      res.status(400).json({ error: "amount must be greater than zero." });
      return;
    }

    const stripe = await getUncachableStripeClient();
    let customerId: string | undefined;

    if (body.customerEmail) {
      const existing = await stripe.customers.list({ email: body.customerEmail, limit: 1 });
      if (existing.data.length > 0) {
        customerId = existing.data[0].id;
      } else {
        const customer = await stripe.customers.create({
          email: body.customerEmail,
          name: body.customerName,
          metadata: { payer_profile_id: body.payerProfileId ?? "" },
        });
        customerId = customer.id;
      }
    }

    const pi = await stripe.paymentIntents.create({
      amount: amountCents,
      currency: "brl",
      payment_method_types: ["pix"],
      customer: customerId,
      metadata: {
        contract_id: contractId,
        payer_profile_id: body.payerProfileId ?? "",
        payee_profile_id: body.payeeProfileId ?? "",
      },
    });

    const confirmed = await stripe.paymentIntents.confirm(pi.id, {
      payment_method_data: { type: "pix" },
    } as Parameters<typeof stripe.paymentIntents.confirm>[1]);

    const pixAction = (confirmed.next_action as Record<string, unknown> | null)
      ?.pix_display_qr_code as
      | { data?: string; image_url_png?: string; image_url_svg?: string; expires_at?: number }
      | undefined;

    await upsertStripePaymentLink({
      contract_id: contractId,
      stripe_payment_intent_id: confirmed.id,
      amount_cents: amountCents,
      currency: "brl",
      status: confirmed.status,
      customer_id: customerId ?? null,
      payer_profile_id: body.payerProfileId ?? null,
      payee_profile_id: body.payeeProfileId ?? null,
      metadata: confirmed.metadata,
    });

    res.status(201).json({
      paymentIntentId: confirmed.id,
      clientSecret: confirmed.client_secret,
      pixCode: pixAction?.data ?? null,
      pixQrImageUrl: pixAction?.image_url_png ?? pixAction?.image_url_svg ?? null,
      expiresAt: pixAction?.expires_at ?? null,
    });
  } catch (error) {
    res.status(500).json({
      error: error instanceof Error ? error.message : "Failed to create Pix PaymentIntent.",
    });
  }
});

export default router;