# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface StripeConfigResponse {
  configured: boolean;
  /** @nullable */
  publishableKey: string | null;
  /** @nullable */
  source?: string | null;
}

export type CreatePaymentIntentRequestMetadata = {[key: string]: string};

export interface CreatePaymentIntentRequest {
  contractId: string;
  amount?: number;
  amountCents?: number;
  currency?: string;
  customerEmail?: string;
  customerName?: string;
  payerProfileId?: string;
  payeeProfileId?: string;
  metadata?: CreatePaymentIntentRequestMetadata;
}

export interface CreatePaymentIntentResponse {
  paymentIntentId: string;
  /** @nullable */
  clientSecret?: string | null;
  status: string;
  amountCents?: number;
  currency?: string;
}

export type PaymentIntentStatusResponseMetadata = { [key: string]: unknown };

export interface PaymentIntentStatusResponse {
  paymentIntentId: string;
  status: string;
  amountCents: number;
  currency: string;
  metadata?: PaymentIntentStatusResponseMetadata;
}

export interface CancelPaymentIntentResponse {
  paymentIntentId: string;
  status: string;
}

export type StripePaymentLinkMetadata = { [key: string]: unknown };

export interface StripePaymentLink {
  /** @nullable */
  id?: string | null;
  contract_id: string;
  stripe_payment_intent_id: string;
  amount_cents: number;
  currency: string;
  status: string;
  /** @nullable */
  customer_id?: string | null;
  /** @nullable */
  payer_profile_id?: string | null;
  /** @nullable */
  payee_profile_id?: string | null;
  metadata: StripePaymentLinkMetadata;
  /** @nullable */
  created_at?: string | null;
  /** @nullable */
  updated_at?: string | null;
}

export interface ContractPaymentsResponse {
  data: StripePaymentLink[];
}

export interface CreateSetupIntentRequest {
  customerEmail?: string;
  customerName?: string;
  payerProfileId?: string;
}

export interface CreateSetupIntentResponse {
  setupIntentId: string;
  /** @nullable */
  clientSecret?: string | null;
}

export interface CreatePixIntentRequest {
  contractId: string;
  amount?: number;
  amountCents?: number;
  customerEmail?: string;
  customerName?: string;
  payerProfileId?: string;
  payeeProfileId?: string;
}

export interface CreatePixIntentResponse {
  paymentIntentId: string;
  /** @nullable */
  clientSecret?: string | null;
  /** @nullable */
  pixCode?: string | null;
  /** @nullable */
  pixQrImageUrl?: string | null;
  /** @nullable */
  expiresAt?: number | null;
}

export type GetStripeConfig503 = StripeConfigResponse & {
  error?: string;
};