# lib/api-zod/src/generated/types/cancelPaymentIntentResponse.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface CancelPaymentIntentResponse {
  paymentIntentId: string;
  status: string;
}