# lib/api-zod/src/generated/types/contractPaymentsResponse.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { StripePaymentLink } from './stripePaymentLink';

export interface ContractPaymentsResponse {
  data: StripePaymentLink[];
}