# lib/api-zod/src/generated/types/createPaymentIntentRequest.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { CreatePaymentIntentRequestMetadata } from './createPaymentIntentRequestMetadata';

export interface CreatePaymentIntentRequest {
  contractId: string;
  amount?: number;
  amountCents?: number;
  currency?: string;
  customerEmail?: string;
  customerName?: string;
  payerProfileId?: string;
  payeeProfileId?: string;
  metadata?: CreatePaymentIntentRequestMetadata;
}