# lib/api-zod/src/generated/types/createPaymentIntentResponse.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface CreatePaymentIntentResponse {
  paymentIntentId: string;
  /** @nullable */
  clientSecret?: string | null;
  status: string;
  amountCents?: number;
  currency?: string;
}