# lib/api-zod/src/generated/types/createPixIntentRequest.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface CreatePixIntentRequest {
  contractId: string;
  amount?: number;
  amountCents?: number;
  customerEmail?: string;
  customerName?: string;
  payerProfileId?: string;
  payeeProfileId?: string;
}