# lib/api-zod/src/generated/types/createPixIntentResponse.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface CreatePixIntentResponse {
  paymentIntentId: string;
  /** @nullable */
  clientSecret?: string | null;
  /** @nullable */
  pixCode?: string | null;
  /** @nullable */
  pixQrImageUrl?: string | null;
  /** @nullable */
  expiresAt?: number | null;
}