# lib/api-zod/src/generated/types/createSetupIntentRequest.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface CreateSetupIntentRequest {
  customerEmail?: string;
  customerName?: string;
  payerProfileId?: string;
}