# lib/api-zod/src/generated/types/createSetupIntentResponse.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface CreateSetupIntentResponse {
  setupIntentId: string;
  /** @nullable */
  clientSecret?: string | null;
}