# lib/api-zod/src/generated/types/paymentIntentStatusResponse.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { PaymentIntentStatusResponseMetadata } from './paymentIntentStatusResponseMetadata';

export interface PaymentIntentStatusResponse {
  paymentIntentId: string;
  status: string;
  amountCents: number;
  currency: string;
  metadata?: PaymentIntentStatusResponseMetadata;
}