# lib/api-zod/src/generated/types/stripeConfigResponse.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface StripeConfigResponse {
  configured: boolean;
  /** @nullable */
  publishableKey: string | null;
  /** @nullable */
  source?: string | null;
}