# lib/api-zod/src/generated/types/stripePaymentLink.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { StripePaymentLinkMetadata } from './stripePaymentLinkMetadata';

export interface StripePaymentLink {
  /** @nullable */
  id?: string | null;
  contract_id: string;
  stripe_payment_intent_id: string;
  amount_cents: number;
  currency: string;
  status: string;
  /** @nullable */
  customer_id?: string | null;
  /** @nullable */
  payer_profile_id?: string | null;
  /** @nullable */
  payee_profile_id?: string | null;
  metadata: StripePaymentLinkMetadata;
  /** @nullable */
  created_at?: string | null;
  /** @nullable */
  updated_at?: string | null;
}